+++ src/target.rs
@@ -56,7 +56,7 @@ impl Target for StderrTarget {
 /// ```
 pub struct FileTarget {
     /// file handle
-    pub file: ReentrantMutex<RefCell<std::fs::File>>,
+    pub file: Mutex<std::fs::File>,
 }
 
 impl FileTarget {
@@ -66,17 +66,19 @@ impl FileTarget {
             std::fs::create_dir_all(dir)?;
         }
 
-        Ok(Self {file: ReentrantMutex::new(RefCell::new(std::fs::OpenOptions::new().create(true).append(true).open(path)?))})
+        Ok(Self {file: Mutex::new(std::fs::OpenOptions::new().create(true).append(true).open(path)?)})
     }
 }
 
 impl Target for FileTarget {
     #[inline]
     fn write(&self, buf: &[u8]) {
-        let file = self.file.lock();
-        match file.borrow_mut().write_all(buf) {
-            Ok(_) => {}
-            Err(err) => { let _ = std::io::stderr().write_all(format!("Error: {}", err).as_bytes()); }
+        match self.file.lock() {
+            Ok(mut obj) => match obj.write_all(buf) {
+                Ok(_) => {}
+                Err(err) => { eprintln!("Error: {}", err); }
+            }
+            Err(err) => { eprintln!("Error: {}", err); }
         };
     }
 }